refactor(platform): split window and render by backend
