+++ Cargo.toml
@@ -18,7 +18,7 @@ crate-type = ["lib"]
 default = ["webgl"]
 # WebGL/WebGPU via a surfman connection. Off for the handheld build: surfman can't
 # initialize on the EGL 1.4 Mali blobs (Knulli/muOS/ROCKNIX). See
-# `src/platform/render.rs::create_surfman_connection`.
+# `src/platform/render/sdl.rs::create_surfman_connection`.
 #
 # `servo/webgl` is what compiles the WebGL thread and the script-side API into the
 # engine; without it the handheld build drops both instead of shipping a WebGL that
@@ -35,7 +35,7 @@ sdl2-static-link = ["sdl2/static-link"]
 # .github/workflows/*, so local `cargo build --release` stays fast.
 #
 # `panic = "abort"` must stay out of here: it makes `catch_unwind` inert and breaks the
-# surfman EGL-1.4 probe (`src/platform/render.rs`). Only the webgl-off handheld CI build
+# surfman EGL-1.4 probe (`src/platform/render/sdl.rs`). Only the webgl-off handheld CI build
 # sets it, where there is no probe.
 [profile.release]
 # `strip` is for local and Android builds, which have no strip step of their own and
